+++ node/src/reactor/joiner.rs
@@ -54,7 +54,7 @@ use crate::{
         requests::{
             BeginGossipRequest, ChainspecLoaderRequest, ConsensusRequest, ContractRuntimeRequest,
             FetcherRequest, MarkBlockCompletedRequest, MetricsRequest, NetworkInfoRequest,
-            NetworkRequest, RestRequest, StorageRequest,
+            NetworkRequest, NodeStateRequest, RestRequest, StorageRequest,
         },
         EffectBuilder, EffectExt, Effects,
     },
@@ -68,7 +68,7 @@ use crate::{
     },
     types::{
         Block, BlockAndDeploys, BlockHeader, BlockHeaderWithMetadata, BlockHeadersBatch,
-        BlockSignatures, BlockWithMetadata, Deploy, ExitCode, FinalizedApprovalsWithId, NodeState,
+        BlockSignatures, BlockWithMetadata, Deploy, ExitCode, FinalizedApprovalsWithId,
     },
     utils::WithDir,
     NodeRng,
@@ -97,6 +97,8 @@ pub(crate) enum JoinerEvent {
     ChainspecLoader(#[serde(skip_serializing)] chainspec_loader::Event),
     #[from]
     ChainspecLoaderRequest(#[serde(skip_serializing)] ChainspecLoaderRequest),
+    #[from]
+    ChainSynchronizerRequest(#[serde(skip_serializing)] NodeStateRequest),
     #[from]
     NetworkInfoRequest(#[serde(skip_serializing)] NetworkInfoRequest),
     #[from]
@@ -230,6 +232,7 @@ impl ReactorEvent for JoinerEvent {
             JoinerEvent::MetricsRequest(_) => "MetricsRequest",
             JoinerEvent::ChainspecLoader(_) => "ChainspecLoader",
             JoinerEvent::ChainspecLoaderRequest(_) => "ChainspecLoaderRequest",
+            JoinerEvent::ChainSynchronizerRequest(_) => "ChainSynchronizerRequest",
             JoinerEvent::NetworkInfoRequest(_) => "NetworkInfoRequest",
             JoinerEvent::BlockFetcher(_) => "BlockFetcher",
             JoinerEvent::BlockByHeightFetcher(_) => "BlockByHeightFetcher",
@@ -331,6 +334,9 @@ impl Display for JoinerEvent {
             JoinerEvent::ChainspecLoader(event) => write!(f, "chainspec loader: {}", event),
             JoinerEvent::ChainspecLoaderRequest(req) => {
                 write!(f, "chainspec loader request: {}", req)
+            }
+            JoinerEvent::ChainSynchronizerRequest(req) => {
+                write!(f, "chain synchronizer request: {}", req)
             }
             JoinerEvent::StorageRequest(req) => write!(f, "storage request: {}", req),
             JoinerEvent::MarkBlockCompletedRequest(req) => {
@@ -564,17 +570,11 @@ impl reactor::Reactor for Reactor {
             .protocol_config
             .verifiable_chunked_hash_activation;
         let protocol_version = &chainspec_loader.chainspec().protocol_config.version;
-        let node_state = if config.node.sync_to_genesis {
-            NodeState::SyncingToGenesis
-        } else {
-            NodeState::FastSyncing
-        };
         let rest_server = RestServer::new(
             config.rest_server.clone(),
             effect_builder,
             *protocol_version,
             node_startup_instant,
-            node_state,
         )?;
 
         let event_stream_server = EventStreamServer::new(
@@ -890,6 +890,11 @@ impl reactor::Reactor for Reactor {
                 effect_builder,
                 rng,
                 JoinerEvent::ChainspecLoader(req.into()),
+            ),
+            JoinerEvent::ChainSynchronizerRequest(req) => self.dispatch_event(
+                effect_builder,
+                rng,
+                JoinerEvent::ChainSynchronizer(req.into()),
             ),
             JoinerEvent::NetworkInfoRequest(req) => {
                 let event = JoinerEvent::SmallNetwork(small_network::Event::from(req));